// frontend/travel-itinerary-planner/src/app/shared/models/auth.model.ts
export interface LoginRequest {

  email: string;

  password: string;

}

export interface RegisterRequest {

  fullName: string;

  email: string;

  password: string;

}

export interface AuthResponse {

  userId: number;

  fullName: string;

  email: string;

  role: string;

  token: string;

}